+++ src/v0/destinations/ga4_v2/transform.ts
@@ -1,8 +1,4 @@
-import {
-  InstrumentationError,
-  isDefinedAndNotNull,
-  RudderStackEvent,
-} from '@rudderstack/integrations-lib';
+import { InstrumentationError, isDefinedAndNotNull } from '@rudderstack/integrations-lib';
 import { ProcessorTransformationRequest } from '../../../types';
 import { handleCustomMappings } from './customMappingsHandler';
 import { processEvents as ga4Process } from '../ga4/transform';
@@ -23,7 +19,7 @@ export function process(event: ProcessorTransformationRequest) {
     Config.apiSecret = configDetails.MEASUREMENT_PROTOCOL_SECRET;
   }
 
-  const eventPayload = message as RudderStackEvent;
+  const eventPayload = message as any;
 
   if (!eventPayload.type) {
     throw new InstrumentationError('Message Type is not present. Aborting message.');
