+++ src/v0/destinations/marketo_static_list/transform.js
@@ -13,7 +13,7 @@ const { getIds, validateMessageType } = require('./util');
 const { getDestinationExternalID, defaultRequestConfig } = require('../../util');
 const { formatConfig, MAX_LEAD_IDS_SIZE } = require('./config');
 const Cache = require('../../util/cache');
-const { getAuthToken } = require('../marketo/transform');
+const { getAuthToken } = require('../marketo/util');
 const { processRecordInputs } = require('./transformV2');
 
 const authCache = new Cache(AUTH_CACHE_TTL); // 1 hr
@@ -90,7 +90,7 @@ const processEvent = (event) => {
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const process = async (event, _processParams) => {
-  const token = await getAuthToken(formatConfig(event.destination), event.metadata);
+  const token = await getAuthToken(authCache, formatConfig(event.destination), event.metadata);
   if (!token) {
     throw new UnauthorizedError('Authorization failed');
   }
@@ -105,7 +105,7 @@ const processRouterDest = async (inputs, reqMetadata) => {
 
   const { destination, metadata } = inputs[0];
   try {
-    const token = await getAuthToken(formatConfig(destination), metadata);
+    const token = await getAuthToken(authCache, formatConfig(destination), metadata, authCache);
     if (!token) {
       throw new UnauthorizedError('Could not retrieve authorisation token');
     }
